Add tests for both 256 and 512 vectors
