+++ test/hotspot/jtreg/compiler/vectorapi/VectorReinterpretTest512.java
@@ -30,28 +30,16 @@ import jdk.incubator.vector.VectorSpecies;
 /*
  * @test
  * @bug 8259775
- * @summary Incorrect code-gen for VectorReinterpret operation
+ * @summary test code-gen for VectorReinterpret operation for 512-bit vectors
  * @modules jdk.incubator.vector
- * @run main compiler.vectorapi.VectorReinterpretTest
+ * @run main compiler.vectorapi.VectorReinterpretTest512
  */
 
-public class VectorReinterpretTest {
-    static final VectorSpecies<Byte> SPECIES_128 = ByteVector.SPECIES_128;
+public class VectorReinterpretTest512 {
     static final VectorSpecies<Byte> SPECIES_256 = ByteVector.SPECIES_256;
     static final VectorSpecies<Byte> SPECIES_512 = ByteVector.SPECIES_512;
 
     static byte[] a = new byte[64];
-
-    private static void test256_128_256() {
-        ByteVector av = ByteVector.fromArray(SPECIES_256, a, 0);
-        ByteVector bv = (ByteVector)av.reinterpretShape(SPECIES_128, 0);
-        ByteVector cv = (ByteVector)bv.reinterpretShape(SPECIES_256, 0);
-
-        if (bv.reduceLanes(VectorOperators.ADD) != 16 ||
-            cv.reduceLanes(VectorOperators.ADD) != 16) {
-            throw new Error("Failed");
-        }
-    }
 
     private static void test512_256_512() {
         ByteVector av = ByteVector.fromArray(SPECIES_512, a, 0);
@@ -70,7 +58,6 @@ public class VectorReinterpretTest {
         }
 
         for (int i = 0; i < 100000; i++) {
-            test256_128_256();
             test512_256_512();
         }
     }
